requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Django==4.1.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
future==0.18.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.1
python-dotenv==0.21.0
python-snappy==0.6.1
pytz==2022.5
serializers==0.2.4
six==1.16.0
sqlparse==0.2.4
Werkzeug==2.2.2
zipp==3.10.0
